% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_gsem.R
\name{caseri}
\alias{caseri}
\title{Filter epidemic data}
\usage{
caseri(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times}
}
\value{
matrix: infection times, removal times
}
\description{
Keep cases only.
}
