% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_gsem.R
\name{impute_gsem}
\alias{impute_gsem}
\title{Missing epidemic data}
\usage{
impute_gsem(epi, p, q = 1)
}
\arguments{
\item{epi}{infection and removal times}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}
}
\description{
Impute NAs for infection and removal times
}
