% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{likelihood_complete_gsem}
\alias{likelihood_complete_gsem}
\title{Likelihood for complete general stochastic epidemic}
\usage{
likelihood_complete_gsem(rates, r, i, N)
}
\arguments{
\item{rates}{numeric vector: rates (beta, gamma)}

\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
negative log likelihood
}
\description{
Compute exact likelihood for complete epidemic observations.
}
