% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete.R
\name{mle_gamma}
\alias{mle_gamma}
\title{MLE for Removal Rate}
\usage{
mle_gamma(r, i)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}
}
\value{
named vector (estimate, sample size)
}
\description{
Compute the maximum likelihood estimate for the removal rate.
}
