% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_complete_gsem.R
\name{mle_gsem}
\alias{mle_gsem}
\title{MLE for complete general stochastic epidemic model}
\usage{
mle_gsem(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
MLE for (beta, gamma)
}
\description{
Compute MLE for complete epidemic observations.
}
