% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete.R
\name{pbla_incomplete_beta}
\alias{pbla_incomplete_beta}
\title{PBLA for incomplete general stochastic epidemic model}
\usage{
pbla_incomplete_beta(beta, r, i, N, A = 1)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: of infection times}

\item{N}{integer: population size}

\item{A}{integer: plausible patient zeros}

\item{rates}{numeric vector: beta (gamma is plug-in MLE of exponential rate)}
}
\value{
negative log likelihood
}
\description{
Compute approximate likelihood for incomplete epidemic observations.
}
