% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_incomplete_gsem.R
\name{pbla_incomplete_gsem}
\alias{pbla_incomplete_gsem}
\title{PBLA for incomplete general stochastic epidemic model}
\usage{
pbla_incomplete_gsem(rates, r, i, N, A = 1)
}
\arguments{
\item{rates}{numeric vector: rates (beta, gamma)}

\item{r}{numeric vector: removal times}

\item{i}{numeric vector: of infection times}

\item{N}{integer: population size}

\item{A}{integer: plausible patient zeros}
}
\value{
negative log likelihood
}
\description{
Compute approximate likelihood for incomplete epidemic observations.
}
