% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_partial_gsem.R
\name{pbla_partial_gsem}
\alias{pbla_partial_gsem}
\title{PBLA for partial general stochastic epidemic model}
\usage{
pbla_partial_gsem(rates, r, N, m = 1, A = 1, lag = 0)
}
\arguments{
\item{rates}{numeric vector: rates (beta, gamma)}

\item{r}{numeric vector: removal times}

\item{N}{integer: population size}

\item{m}{integer: positive shape}

\item{A}{integer: plausible patient zeros}

\item{lag}{numeric: fixed lag}
}
\value{
negative log likelihood
}
\description{
Compute approximate likelihood for partial epidemic observations.
}
