% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sort_gsem}
\alias{sort_gsem}
\title{Sort epidemic data}
\usage{
sort_gsem(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times}
}
\value{
matrix: infection times, removal times
}
\description{
Sort matrix by increasing removal times.
}
