% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_gsem.R
\name{sortri}
\alias{sortri}
\title{Sort epidemic data}
\usage{
sortri(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times}
}
\value{
matrix: infection times, removal times
}
\description{
Sort matrix by increasing removal times.
}
