% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsem.R
\name{rgsem}
\alias{rgsem}
\title{Simulate general stochastic epidemic model}
\usage{
rgsem(beta, gamma, N, m = 1)
}
\arguments{
\item{N}{integer: population size}

\item{m}{integer: positive shape}

\item{rates}{numeric vector: rates (beta, gamma)}
}
\value{
matrix: infection times, removal times
}
\description{
Draw infectious periods for general stochastic epidemic.
}
